requirements.txt
addict==2.1.3
alabaster==0.7.10
altgraph==0.15
anaconda-client==1.6.5
anaconda-navigator==1.6.8
anaconda-project==0.8.0
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.0.0
bokeh==0.12.7
boto==2.48.0
Bottleneck==1.2.1
CacheControl==0.12.3
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.2
conda==4.3.27
conda-build==3.0.22
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.0.3
cx-Freeze==5.1.1
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.2
dataclasses==0.6
datashape==0.5.4
decorator==4.1.2
defusedxml==0.5.0
distlib==0.2.5
distributed==1.18.3
dj-database-url==0.5.0
Django==2.0.6
django-ckeditor==5.4.0
django-filter==1.1.0
django-heroku==0.3.1
django-jet==1.0.7
django-js-asset==1.1.0
django-taggit==0.22.2
djangorestframework==3.8.2
djongo==1.2.29
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
facebook-sdk==2.0.0
fastcache==1.0.2
fbchat==1.3.9
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
gevent==1.2.2
glob2==0.5
gmail==0.6.3
greenlet==0.4.12
gunicorn==19.8.1
h5py==2.7.0
heapdict==1.0.0
html5lib==0.999999999
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==3.8.0
macholib==1.9
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
menuinst==1.4.8
mistune==0.7.4
mongoengine==0.15.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e
numexpr==2.6.2
numpy==1.13.1
numpydoc==0.7.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
opencv-contrib-python==3.4.1.15
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pefile==2017.11.5
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
ply==3.10
progress==1.3
prompt-toolkit==1.0.15
psutil==5.2.2
psycopg2==2.7.5
py==1.4.34
py2exe==0.9.2.2
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
pygame==1.9.3
Pygments==2.2.0
PyInstaller==3.3.1
PyJWT==1.6.4
pylint==1.7.2
pymongo==3.6.1
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.4
PyWavelets==0.5.2
pywin32==221
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
requests-oauthlib==1.0.0
rope==0.10.5
ruamel-yaml==0.11.14
schedule==0.5.0
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
SQLAlchemy==1.1.13
sqlparse==0.2.4
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.1
widgetsnbextension==3.0.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==0.9.8
xlwings==0.11.4
xlwt==1.3.0
youtube-dl==2017.12.31
zict==0.1.2
